+++ src/pages/messages/components/NviCandidatesList.tsx
@@ -1,9 +1,10 @@
-import { Box, List, Typography } from '@mui/material';
+import { Box, List, MenuItem, Select, Typography } from '@mui/material';
 import { UseQueryResult } from '@tanstack/react-query';
 import { Dispatch, SetStateAction } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { useTranslation } from 'react-i18next';
-import { FetchNviCandidatesParams } from '../../../api/searchApi';
+import { useHistory } from 'react-router-dom';
+import { FetchNviCandidatesParams, NviCandidatesSearchParam } from '../../../api/searchApi';
 import { CuratorSelector } from '../../../components/CuratorSelector';
 import { ErrorBoundary } from '../../../components/ErrorBoundary';
 import { ListPagination } from '../../../components/ListPagination';
@@ -12,6 +13,7 @@ import { SearchForm } from '../../../components/SearchForm';
 import { NviCandidateSearchResponse } from '../../../types/nvi.types';
 import { RoleName } from '../../../types/user.types';
 import { dataTestId } from '../../../utils/dataTestIds';
+import { getNviYearFilterValues } from '../../../utils/nviHelpers';
 import { NviCandidateListItem } from './NviCandidateListItem';
 
 interface NviCandidatesListProps {
@@ -24,6 +26,8 @@ interface NviCandidatesListProps {
   helmetTitle: string;
 }
 
+const nviYearFilterValues = getNviYearFilterValues();
+
 export const NviCandidatesList = ({
   nviCandidatesQuery,
   nviQueryParams,
@@ -34,6 +38,12 @@ export const NviCandidatesList = ({
   helmetTitle,
 }: NviCandidatesListProps) => {
   const { t } = useTranslation();
+  const history = useHistory();
+
+  const searchParams = new URLSearchParams(history.location.search);
+
+  const nviYearParam = searchParams.get(NviCandidatesSearchParam.Year);
+  const nviYearFilter = nviYearParam ? +nviYearParam : nviYearFilterValues[1];
 
   return (
     <section>
@@ -43,7 +53,25 @@ export const NviCandidatesList = ({
 
       <Box sx={{ mb: '1rem', mx: { xs: '0.5rem', md: 0, display: 'grid', gap: '1rem' } }}>
         <SearchForm placeholder={t('tasks.search_placeholder')} />
-        <CuratorSelector roleFilter={[RoleName.NviCurator]} sx={{ maxWidth: '20rem' }} />
+
+        <Box sx={{ display: 'flex', gap: '1rem' }}>
+          <CuratorSelector roleFilter={[RoleName.NviCurator]} sx={{ maxWidth: '20rem' }} />
+          <Select
+            size="small"
+            inputProps={{ 'aria-label': t('common.year') }}
+            value={nviYearFilter}
+            onChange={(event) => {
+              searchParams.set(NviCandidatesSearchParam.Year, event.target.value.toString());
+              history.push({ search: searchParams.toString() });
+            }}
+            sx={{ width: 'fit-content', alignSelf: 'center', mb: '0.5rem' }}>
+            {nviYearFilterValues.map((year) => (
+              <MenuItem key={year} value={year}>
+                {year}
+              </MenuItem>
+            ))}
+          </Select>
+        </Box>
       </Box>
 
       {nviCandidatesQuery.isPending ? (
